Update dependencies and enable fatal warnings (#1885)

+++ eclair-core/src/main/scala/fr/acinq/eclair/blockchain/watchdogs/BlockchainWatchdog.scala
@@ -93,7 +93,7 @@ object BlockchainWatchdog {
             } else {
               context.log.info("{}: we are {} blocks late", source, missingBlocks)
             }
-            Metrics.BitcoinBlocksSkew.withTag(Tags.Source, source).update(missingBlocks)
+            Metrics.BitcoinBlocksSkew.withTag(Tags.Source, source).update(missingBlocks.toDouble)
             Behaviors.same
         }
       }
